feat:全文索引查询功能实现,更新后老索引未失效错误待排查

+++ myblog-server/src/main/java/com/yikolemon/pojo/Like.java
@@ -12,7 +12,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 @AllArgsConstructor                //有参构造
 public class Like implements Serializable {
 
-    private int like;
+    private long like;
 
     private long blogId;
